fix: calc difference value

+++ src/Halle.Domain/Entities/Stock/Stock.cs
@@ -4,6 +4,6 @@
     {
         public string Ticker { get; set; } = null!;
         public int Qtd { get; set; }
-        public double Goal { get; set; }
+        public decimal Goal { get; set; }
     }
 }
